+++ packages/util/src/hex/toU8a.ts
@@ -28,37 +28,25 @@ export function hexToU8a (_value?: HexString | string | null, bitLength = -1): U
 
   const value = hexStripPrefix(_value).toLowerCase();
   const valLength = value.length / 2;
-  const resultLength = Math.ceil(
+  const endLength = Math.ceil(
     bitLength === -1
       ? valLength
       : bitLength / 8
   );
-  const result = new Uint8Array(resultLength);
-  const offset = resultLength > valLength
-    ? resultLength - valLength
+  const result = new Uint8Array(endLength);
+  const offset = endLength > valLength
+    ? endLength - valLength
     : 0;
+  const dv = new DataView(result.buffer, offset);
+  const mod = (endLength - offset) % 2;
+  const length = endLength - offset - mod;
 
-  // With offsets, it becomes slightly trickier since we need to ensure the U16
-  // is at the correct alignment as well. Due to this, we just focus on the simple
-  // case where we do no bitlength conversions (which is the most-taken path)
-  if (offset) {
-    for (let i = 0; i < resultLength; i++) {
-      result[i + offset] = HEX_TO_U8[value.substr(i * 2, 2)];
-    }
-  } else {
-    // NOTE: It is only _slightly_ more optimal to use HEX_TO_U16 here as
-    // well. The overhead does not seem to be in the lookup map iteration,
-    // even with 640k inputs
-    const mod = resultLength % 2;
-    const u16 = new Uint16Array(result.buffer, 0, (resultLength - mod) / 2);
+  for (let i = 0; i < length; i += 2) {
+    dv.setUint16(i, HEX_TO_U16[value.substr(i * 2, 4)]);
+  }
 
-    for (let i = 0; i < u16.length; i++) {
-      u16[i] = HEX_TO_U16[value.substr(i * 4, 4)];
-    }
-
-    if (mod) {
-      result[resultLength - 1] = HEX_TO_U8[value.substr(value.length - 2, 2)];
-    }
+  if (mod) {
+    dv.setUint8(length, HEX_TO_U8[value.substr(value.length - 2, 2)]);
   }
 
   return result;